+++ packages/schemas/src/consts/index.ts
@@ -1,3 +1,2 @@
 export * from './system.js';
 export * from './oidc.js';
-export * from './storage.js';
