--- packages/schemas/src/consts/storage.ts
@@ -1,13 +1,0 @@
-export const maxUploadFileSize = 8 * 1024 * 1024; // 8MB
-
-// Reference: https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types/Common_types
-export const allowUploadMimeTypes = [
-  'image/jpeg',
-  'image/png',
-  'image/gif',
-  'image/vnd.microsoft.icon',
-  'image/svg+xml',
-  'image/tiff',
-  'image/webp',
-  'image/bmp',
-];
