+++ packages/schemas/src/types/user-assets.ts
@@ -1,13 +1,33 @@
-import type { allowUploadMimeTypes } from '../index.js';
+import { z } from 'zod';
 
-export type AllowedUploadMimeType = (typeof allowUploadMimeTypes)[number];
+export const maxUploadFileSize = 8 * 1024 * 1024; // 8MB
 
-export type UserAssetsServiceStatusResponse = {
-  status: 'ready' | 'not_configured';
-  allowUploadMimeTypes?: string[];
-  maxUploadFileSize?: number;
-};
+// Reference: https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types/Common_types
+export const allowUploadMimeTypes = [
+  'image/jpeg',
+  'image/png',
+  'image/gif',
+  'image/vnd.microsoft.icon',
+  'image/svg+xml',
+  'image/tiff',
+  'image/webp',
+  'image/bmp',
+] as const;
 
-export type UserAssetsResponse = {
-  url: string;
-};
+const allowUploadMimeTypeGuard = z.enum(allowUploadMimeTypes);
+
+export type AllowedUploadMimeType = z.infer<typeof allowUploadMimeTypeGuard>;
+
+export const userAssetsServiceStatusGuard = z.object({
+  status: z.union([z.literal('ready'), z.literal('not_configured')]),
+  allowUploadMimeTypes: z.array(allowUploadMimeTypeGuard).optional(),
+  maxUploadFileSize: z.number().optional(),
+});
+
+export type UserAssetsServiceStatusResponse = z.infer<typeof userAssetsServiceStatusGuard>;
+
+export const userAssetsGuard = z.object({
+  url: z.string(),
+});
+
+export type UserAssetsResponse = z.infer<typeof userAssetsGuard>;
